functions/process_ocr/requirements.txt
google-cloud-storage
google-cloud-documentai
google-cloud-bigquery
google-cloud-pubsub
functions-framework